Change order of return values from getyx - fixes display of historical CPU.
In this commit:
http://cvs.savannah.gnu.org/viewvc/ocaml-tmk/functions.c?root=ocaml-tmk&r1=1.2&r2=1.3
Sam changed (fixed) the order that the coordinates are returned
from the ncurses 'getyx' call.  Since this commit was made 3 years
ago it should be safe to assume everyone is using the new ocaml-ncurses
by now.

This meant the historical %CPU was being displayed off the bottom of
the screen instead of on the top row at the right.  Now it is displayed
in the correct place.