Fix pad function to work for negative widths (RHBZ#634435).
The pad function is supposed to pad a string to a width.  Under some
circumstances, the width parameter could be negative resulting in a
call to String.sub such as:

  String.sub str 0 (-1)

which caused an exception to be thrown and not caught, causing
virt-top to exit.  Fix the pad function to return an empty string if
width <= 0 instead of throwing an exception.

See also:
https://bugzilla.redhat.com/show_bug.cgi?id=634435